added payment tests

+++ playground/core/services/interfaces/currency_converter_interface.py
@@ -2,5 +2,5 @@ from typing import Protocol
 
 
 class ICurrencyConverter(Protocol):
-    def convert(self, from_currency: str, to_currency: str, amount: int) -> float:
+    def convert(self, from_currency: str, to_currency: str, amount: float) -> float:
         pass
